+++ junos/resource_system_syslog_file.go
@@ -6,18 +6,14 @@ import (
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	jdecode "github.com/jeremmfr/junosdecode"
 )
 
-type syslogHostOptions struct {
+type syslogFileOptions struct {
 	allowDuplicates             bool
-	excludeHostname             bool
 	explicitPriority            bool
-	port                        int
-	host                        string
-	facilityOverride            string
-	logPrefix                   string
+	filename                    string
 	match                       string
-	sourceAddress               string
 	anySeverity                 string
 	authorizationSeverity       string
 	changelogSeverity           string
@@ -34,55 +30,33 @@ type syslogHostOptions struct {
 	securitySeverity            string
 	userSeverity                string
 	matchStrings                []string
+	archive                     []map[string]interface{}
 	structuredData              []map[string]interface{}
 }
 
-func resourceSystemSyslogHost() *schema.Resource {
+func resourceSystemSyslogFile() *schema.Resource {
 	return &schema.Resource{
-		Create: resourceSystemSyslogHostCreate,
-		Read:   resourceSystemSyslogHostRead,
-		Update: resourceSystemSyslogHostUpdate,
-		Delete: resourceSystemSyslogHostDelete,
+		Create: resourceSystemSyslogFileCreate,
+		Read:   resourceSystemSyslogFileRead,
+		Update: resourceSystemSyslogFileUpdate,
+		Delete: resourceSystemSyslogFileDelete,
 		Importer: &schema.ResourceImporter{
-			State: resourceSystemSyslogHostImport,
+			State: resourceSystemSyslogFileImport,
 		},
 		Schema: map[string]*schema.Schema{
-			"host": {
+			"filename": {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validateAddress(),
+				ValidateFunc: validateNameObjectJunos(),
 			},
 			"allow_duplicates": {
-				Type:     schema.TypeBool,
-				Optional: true,
-			},
-			"exclude_hostname": {
 				Type:     schema.TypeBool,
 				Optional: true,
 			},
 			"explicit_priority": {
 				Type:     schema.TypeBool,
 				Optional: true,
-			},
-			"facility_override": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ValidateFunc: func(v interface{}, k string) (ws []string, errors []error) {
-					value := v.(string)
-					if !stringInSlice(value, []string{"authorization", "daemon", "ftp", "kernel", "user",
-						"local0", "local1", "local2", "local3", "local4", "local5", "local6", "local7"}) {
-						errors = append(errors, fmt.Errorf(
-							"%q for %q is not a valid facilty", value, k))
-					}
-
-					return
-				},
-			},
-			"log_prefix": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validateNameObjectJunos(),
 			},
 			"match": {
 				Type:     schema.TypeString,
@@ -92,16 +66,6 @@ func resourceSystemSyslogHost() *schema.Resource {
 				Type:     schema.TypeList,
 				Optional: true,
 				Elem:     &schema.Schema{Type: schema.TypeString},
-			},
-			"port": {
-				Type:         schema.TypeInt,
-				Optional:     true,
-				ValidateFunc: validateIntRange(1, 65535),
-			},
-			"source_address": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validateIPFunc(),
 			},
 			"structured_data": {
 				Type:     schema.TypeList,
@@ -191,11 +155,80 @@ func resourceSystemSyslogHost() *schema.Resource {
 				Optional:     true,
 				ValidateFunc: validateSyslogSeverity(),
 			},
+			"archive": {
+				Type:     schema.TypeList,
+				Optional: true,
+				MaxItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"sites": {
+							Type:     schema.TypeList,
+							Optional: true,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"url": {
+										Type:     schema.TypeString,
+										Required: true,
+									},
+									"password": {
+										Type:      schema.TypeString,
+										Optional:  true,
+										Sensitive: true,
+									},
+									"routing_instance": {
+										Type:     schema.TypeString,
+										Optional: true,
+									},
+								},
+							},
+						},
+						"binary_data": {
+							Type:          schema.TypeBool,
+							Optional:      true,
+							ConflictsWith: []string{"archive.0.no_binary_data"},
+						},
+						"no_binary_data": {
+							Type:          schema.TypeBool,
+							Optional:      true,
+							ConflictsWith: []string{"archive.0.binary_data"},
+						},
+						"world_readable": {
+							Type:          schema.TypeBool,
+							Optional:      true,
+							ConflictsWith: []string{"archive.0.no_world_readable"},
+						},
+						"no_world_readable": {
+							Type:          schema.TypeBool,
+							Optional:      true,
+							ConflictsWith: []string{"archive.0.world_readable"},
+						},
+						"files": {
+							Type:         schema.TypeInt,
+							Optional:     true,
+							ValidateFunc: validateIntRange(1, 1000),
+						},
+						"size": {
+							Type:         schema.TypeInt,
+							Optional:     true,
+							ValidateFunc: validateIntRange(65536, 1073741824),
+						},
+						"start_time": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+						"transfer_interval": {
+							Type:         schema.TypeInt,
+							Optional:     true,
+							ValidateFunc: validateIntRange(5, 2880),
+						},
+					},
+				},
+			},
 		},
 	}
 }
 
-func resourceSystemSyslogHostCreate(d *schema.ResourceData, m interface{}) error {
+func resourceSystemSyslogFileCreate(d *schema.ResourceData, m interface{}) error {
 	sess := m.(*Session)
 	jnprSess, err := sess.startNewSession()
 	if err != nil {
@@ -206,43 +239,43 @@ func resourceSystemSyslogHostCreate(d *schema.ResourceData, m interface{}) error
 	if err != nil {
 		return err
 	}
-	syslogHostExists, err := checkSystemSyslogHostExists(d.Get("host").(string), m, jnprSess)
+	syslogFileExists, err := checkSystemSyslogFileExists(d.Get("filename").(string), m, jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
 		return err
 	}
-	if syslogHostExists {
+	if syslogFileExists {
 		sess.configClear(jnprSess)
 
-		return fmt.Errorf("system syslog host %v already exists", d.Get("host").(string))
+		return fmt.Errorf("system syslog file %v already exists", d.Get("filename").(string))
 	}
 
-	err = setSystemSyslogHost(d, m, jnprSess)
+	err = setSystemSyslogFile(d, m, jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
 		return err
 	}
-	err = sess.commitConf("create resource junos_system_syslog_host", jnprSess)
+	err = sess.commitConf("create resource junos_system_syslog_file", jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
 		return err
 	}
-	syslogHostExists, err = checkSystemSyslogHostExists(d.Get("host").(string), m, jnprSess)
+	syslogFileExists, err = checkSystemSyslogFileExists(d.Get("filename").(string), m, jnprSess)
 	if err != nil {
 		return err
 	}
-	if syslogHostExists {
-		d.SetId(d.Get("host").(string))
+	if syslogFileExists {
+		d.SetId(d.Get("filename").(string))
 	} else {
-		return fmt.Errorf("system syslog host %v not exists after commit => check your config", d.Get("host").(string))
+		return fmt.Errorf("system syslog file %v not exists after commit => check your config", d.Get("filename").(string))
 	}
 
-	return resourceSystemSyslogHostRead(d, m)
+	return resourceSystemSyslogFileRead(d, m)
 }
-func resourceSystemSyslogHostRead(d *schema.ResourceData, m interface{}) error {
+func resourceSystemSyslogFileRead(d *schema.ResourceData, m interface{}) error {
 	sess := m.(*Session)
 	mutex.Lock()
 	jnprSess, err := sess.startNewSession()
@@ -252,20 +285,20 @@ func resourceSystemSyslogHostRead(d *schema.ResourceData, m interface{}) error {
 		return err
 	}
 	defer sess.closeSession(jnprSess)
-	syslogHostOptions, err := readSystemSyslogHost(d.Get("host").(string), m, jnprSess)
+	syslogFileOptions, err := readSystemSyslogFile(d.Get("filename").(string), m, jnprSess)
 	mutex.Unlock()
 	if err != nil {
 		return err
 	}
-	if syslogHostOptions.host == "" {
+	if syslogFileOptions.filename == "" {
 		d.SetId("")
 	} else {
-		fillSystemSyslogHostData(d, syslogHostOptions)
+		fillSystemSyslogFileData(d, syslogFileOptions)
 	}
 
 	return nil
 }
-func resourceSystemSyslogHostUpdate(d *schema.ResourceData, m interface{}) error {
+func resourceSystemSyslogFileUpdate(d *schema.ResourceData, m interface{}) error {
 	d.Partial(true)
 	sess := m.(*Session)
 	jnprSess, err := sess.startNewSession()
@@ -277,19 +310,19 @@ func resourceSystemSyslogHostUpdate(d *schema.ResourceData, m interface{}) error
 	if err != nil {
 		return err
 	}
-	err = delSystemSyslogHost(d.Get("host").(string), m, jnprSess)
+	err = delSystemSyslogFile(d.Get("filename").(string), m, jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
 		return err
 	}
-	err = setSystemSyslogHost(d, m, jnprSess)
+	err = setSystemSyslogFile(d, m, jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
 		return err
 	}
-	err = sess.commitConf("update resource junos_system_syslog_host", jnprSess)
+	err = sess.commitConf("update resource junos_system_syslog_file", jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
@@ -297,9 +330,9 @@ func resourceSystemSyslogHostUpdate(d *schema.ResourceData, m interface{}) error
 	}
 	d.Partial(false)
 
-	return resourceSystemSyslogHostRead(d, m)
+	return resourceSystemSyslogFileRead(d, m)
 }
-func resourceSystemSyslogHostDelete(d *schema.ResourceData, m interface{}) error {
+func resourceSystemSyslogFileDelete(d *schema.ResourceData, m interface{}) error {
 	sess := m.(*Session)
 	jnprSess, err := sess.startNewSession()
 	if err != nil {
@@ -310,13 +343,13 @@ func resourceSystemSyslogHostDelete(d *schema.ResourceData, m interface{}) error
 	if err != nil {
 		return err
 	}
-	err = delSystemSyslogHost(d.Get("host").(string), m, jnprSess)
+	err = delSystemSyslogFile(d.Get("filename").(string), m, jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
 		return err
 	}
-	err = sess.commitConf("delete resource junos_system_syslog_host", jnprSess)
+	err = sess.commitConf("delete resource junos_system_syslog_file", jnprSess)
 	if err != nil {
 		sess.configClear(jnprSess)
 
@@ -325,7 +358,7 @@ func resourceSystemSyslogHostDelete(d *schema.ResourceData, m interface{}) error
 
 	return nil
 }
-func resourceSystemSyslogHostImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+func resourceSystemSyslogFileImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
 	sess := m.(*Session)
 	jnprSess, err := sess.startNewSession()
 	if err != nil {
@@ -334,69 +367,55 @@ func resourceSystemSyslogHostImport(d *schema.ResourceData, m interface{}) ([]*s
 	defer sess.closeSession(jnprSess)
 	result := make([]*schema.ResourceData, 1)
 
-	syslogHostExists, err := checkSystemSyslogHostExists(d.Id(), m, jnprSess)
+	syslogFileExists, err := checkSystemSyslogFileExists(d.Id(), m, jnprSess)
 	if err != nil {
 		return nil, err
 	}
-	if !syslogHostExists {
-		return nil, fmt.Errorf("don't find system syslog host with id '%v' (id must be <host>)", d.Id())
+	if !syslogFileExists {
+		return nil, fmt.Errorf("don't find system syslog file with id '%v' (id must be <filename>)", d.Id())
 	}
-	syslogHostOptions, err := readSystemSyslogHost(d.Id(), m, jnprSess)
+	syslogFileOptions, err := readSystemSyslogFile(d.Id(), m, jnprSess)
 	if err != nil {
 		return nil, err
 	}
-	fillSystemSyslogHostData(d, syslogHostOptions)
+	fillSystemSyslogFileData(d, syslogFileOptions)
 
 	result[0] = d
 
 	return result, nil
 }
 
-func checkSystemSyslogHostExists(host string, m interface{}, jnprSess *NetconfObject) (bool, error) {
+func checkSystemSyslogFileExists(filename string, m interface{}, jnprSess *NetconfObject) (bool, error) {
 	sess := m.(*Session)
-	syslogHostConfig, err := sess.command("show configuration"+
-		" system syslog host "+host+" | display set", jnprSess)
+	syslogFileConfig, err := sess.command("show configuration"+
+		" system syslog file "+filename+" | display set", jnprSess)
 	if err != nil {
 		return false, err
 	}
-	if syslogHostConfig == emptyWord {
+	if syslogFileConfig == emptyWord {
 		return false, nil
 	}
 
 	return true, nil
 }
-func setSystemSyslogHost(d *schema.ResourceData, m interface{}, jnprSess *NetconfObject) error {
+
+func setSystemSyslogFile(d *schema.ResourceData, m interface{}, jnprSess *NetconfObject) error {
 	sess := m.(*Session)
 
-	setPrefix := "set system syslog host " + d.Get("host").(string)
+	setPrefix := "set system syslog file " + d.Get("filename").(string)
 	configSet := make([]string, 0)
 
 	if d.Get("allow_duplicates").(bool) {
 		configSet = append(configSet, setPrefix+" allow-duplicates")
 	}
-	if d.Get("exclude_hostname").(bool) {
-		configSet = append(configSet, setPrefix+" exclude-hostname")
-	}
 	if d.Get("explicit_priority").(bool) {
 		configSet = append(configSet, setPrefix+" explicit-priority")
-	}
-	if d.Get("facility_override").(string) != "" {
-		configSet = append(configSet, setPrefix+" facility-override "+d.Get("facility_override").(string))
-	}
-	if d.Get("log_prefix").(string) != "" {
-		configSet = append(configSet, setPrefix+" log-prefix "+d.Get("log_prefix").(string))
 	}
 	if d.Get("match").(string) != "" {
 		configSet = append(configSet, setPrefix+" match \""+d.Get("match").(string)+"\"")
 	}
 	for _, v := range d.Get("match_strings").([]interface{}) {
 		configSet = append(configSet, setPrefix+" match-strings \""+v.(string)+"\"")
-	}
-	if d.Get("port").(int) != 0 {
-		configSet = append(configSet, setPrefix+" port "+strconv.Itoa(d.Get("port").(int)))
-	}
-	if d.Get("source_address").(string) != "" {
-		configSet = append(configSet, setPrefix+" source-address "+d.Get("source_address").(string))
 	}
 	for _, v := range d.Get("structured_data").([]interface{}) {
 		configSet = append(configSet, setPrefix+" structured-data")
@@ -452,6 +471,54 @@ func setSystemSyslogHost(d *schema.ResourceData, m interface{}, jnprSess *Netcon
 	if d.Get("user_severity").(string) != "" {
 		configSet = append(configSet, setPrefix+" user "+d.Get("user_severity").(string))
 	}
+	for _, v := range d.Get("archive").([]interface{}) {
+		setPrefixArchive := setPrefix + " archive"
+		configSet = append(configSet, setPrefixArchive)
+		if v != nil {
+			m := v.(map[string]interface{})
+			for _, v2 := range m["sites"].([]interface{}) {
+				m2 := v2.(map[string]interface{})
+				setPrefixArchiveSite := setPrefixArchive + " archive-sites " + m2["url"].(string)
+				configSet = append(configSet, setPrefixArchiveSite)
+				if m2["password"].(string) != "" {
+					configSet = append(configSet, setPrefixArchiveSite+" password \""+
+						m2["password"].(string)+"\"")
+				}
+				if m2["routing_instance"].(string) != "" {
+					configSet = append(configSet, setPrefixArchiveSite+" routing-instance "+
+						m2["routing_instance"].(string))
+				}
+			}
+			if m["binary_data"].(bool) {
+				configSet = append(configSet, setPrefixArchive+" binary-data")
+			}
+			if m["no_binary_data"].(bool) {
+				configSet = append(configSet, setPrefixArchive+" no-binary-data")
+			}
+			if m["world_readable"].(bool) {
+				configSet = append(configSet, setPrefixArchive+" world-readable")
+			}
+			if m["no_world_readable"].(bool) {
+				configSet = append(configSet, setPrefixArchive+" no-world-readable")
+			}
+			if m["files"].(int) != 0 {
+				configSet = append(configSet, setPrefixArchive+" files "+
+					strconv.Itoa(m["files"].(int)))
+			}
+			if m["size"].(int) != 0 {
+				configSet = append(configSet, setPrefixArchive+" size "+
+					strconv.Itoa(m["size"].(int)))
+			}
+			if m["start_time"].(string) != "" {
+				configSet = append(configSet, setPrefixArchive+" start-time "+
+					m["start_time"].(string))
+			}
+			if m["transfer_interval"].(int) != 0 {
+				configSet = append(configSet, setPrefixArchive+" transfer-interval "+
+					strconv.Itoa(m["transfer_interval"].(int)))
+			}
+		}
+	}
 
 	err := sess.configSet(configSet, jnprSess)
 	if err != nil {
@@ -460,18 +527,19 @@ func setSystemSyslogHost(d *schema.ResourceData, m interface{}, jnprSess *Netcon
 
 	return nil
 }
-func readSystemSyslogHost(host string, m interface{}, jnprSess *NetconfObject) (syslogHostOptions, error) {
-	sess := m.(*Session)
-	var confRead syslogHostOptions
 
-	syslogHostConfig, err := sess.command("show configuration"+
-		" system syslog host "+host+" | display set relative", jnprSess)
+func readSystemSyslogFile(filename string, m interface{}, jnprSess *NetconfObject) (syslogFileOptions, error) {
+	sess := m.(*Session)
+	var confRead syslogFileOptions
+
+	syslogFileConfig, err := sess.command("show configuration"+
+		" system syslog file "+filename+" | display set relative", jnprSess)
 	if err != nil {
 		return confRead, err
 	}
-	if syslogHostConfig != emptyWord {
-		confRead.host = host
-		for _, item := range strings.Split(syslogHostConfig, "\n") {
+	if syslogFileConfig != emptyWord {
+		confRead.filename = filename
+		for _, item := range strings.Split(syslogFileConfig, "\n") {
 			if strings.Contains(item, "<configuration-output>") {
 				continue
 			}
@@ -482,27 +550,13 @@ func readSystemSyslogHost(host string, m interface{}, jnprSess *NetconfObject) (
 			switch {
 			case strings.HasSuffix(itemTrim, "allow-duplicates"):
 				confRead.allowDuplicates = true
-			case strings.HasSuffix(itemTrim, "exclude-hostname"):
-				confRead.excludeHostname = true
 			case strings.HasSuffix(itemTrim, "explicit-priority"):
 				confRead.explicitPriority = true
-			case strings.HasPrefix(itemTrim, "facility-override "):
-				confRead.facilityOverride = strings.TrimPrefix(itemTrim, "facility-override ")
-			case strings.HasPrefix(itemTrim, "log-prefix "):
-				confRead.logPrefix = strings.TrimPrefix(itemTrim, "log-prefix ")
 			case strings.HasPrefix(itemTrim, "match "):
 				confRead.match = strings.Trim(strings.TrimPrefix(itemTrim, "match "), "\"")
 			case strings.HasPrefix(itemTrim, "match-strings "):
 				confRead.matchStrings = append(confRead.matchStrings,
 					strings.Trim(strings.TrimPrefix(itemTrim, "match-strings "), "\""))
-			case strings.HasPrefix(itemTrim, "port "):
-				var err error
-				confRead.port, err = strconv.Atoi(strings.TrimPrefix(itemTrim, "port "))
-				if err != nil {
-					return confRead, err
-				}
-			case strings.HasPrefix(itemTrim, "source-address "):
-				confRead.sourceAddress = strings.TrimPrefix(itemTrim, "source-address ")
 			case strings.HasPrefix(itemTrim, "structured-data"):
 				structuredData := map[string]interface{}{
 					"brief": false,
@@ -542,10 +596,81 @@ func readSystemSyslogHost(host string, m interface{}, jnprSess *NetconfObject) (
 				confRead.securitySeverity = strings.TrimPrefix(itemTrim, "security ")
 			case strings.HasPrefix(itemTrim, "user "):
 				confRead.userSeverity = strings.TrimPrefix(itemTrim, "user ")
+			case strings.HasPrefix(itemTrim, "archive"):
+				archiveM := map[string]interface{}{
+					"sites":             make([]map[string]interface{}, 0),
+					"binary_data":       false,
+					"no_binary_data":    false,
+					"world_readable":    false,
+					"no_world_readable": false,
+					"files":             0,
+					"size":              0,
+					"transfer_interval": 0,
+					"start_time":        "",
+				}
+				if len(confRead.archive) == 1 {
+					archiveM = confRead.archive[0]
+				}
+				switch {
+				case strings.HasSuffix(itemTrim, "archive binary-data"):
+					archiveM["binary_data"] = true
+				case strings.HasSuffix(itemTrim, "archive no-binary-data"):
+					archiveM["no_binary_data"] = true
+				case strings.HasSuffix(itemTrim, "archive world-readable"):
+					archiveM["world_readable"] = true
+				case strings.HasSuffix(itemTrim, "archive no-world-readable"):
+					archiveM["no_world_readable"] = true
+				case strings.HasPrefix(itemTrim, "archive files "):
+					var err error
+					archiveM["files"], err = strconv.Atoi(strings.TrimPrefix(itemTrim, "archive files "))
+					if err != nil {
+						return confRead, err
+					}
+				case strings.HasPrefix(itemTrim, "archive size "):
+					var err error
+					archiveM["size"], err = strconv.Atoi(strings.TrimPrefix(itemTrim, "archive size "))
+					if err != nil {
+						return confRead, err
+					}
+				case strings.HasPrefix(itemTrim, "archive transfer-interval "):
+					var err error
+					archiveM["transfer_interval"], err = strconv.Atoi(strings.TrimPrefix(itemTrim, "archive transfer-interval "))
+					if err != nil {
+						return confRead, err
+					}
+				case strings.HasPrefix(itemTrim, "archive start-time "):
+					archiveM["start_time"] = strings.TrimPrefix(itemTrim, "archive start-time ")
+				case strings.HasPrefix(itemTrim, "archive archive-sites "):
+					itemTrimArchSitesSplit := strings.Split(
+						strings.TrimPrefix(itemTrim, "archive archive-sites "), " ")
+					itemTrimArchSites := strings.TrimPrefix(itemTrim, "archive archive-sites "+itemTrimArchSitesSplit[0]+" ")
+					sitesOptions := map[string]interface{}{
+						"url":              itemTrimArchSitesSplit[0],
+						"password":         "",
+						"routing_instance": "",
+					}
+					sitesOptions, archiveM["sites"] = copyAndRemoveItemMapList("url", false, sitesOptions,
+						archiveM["sites"].([]map[string]interface{}))
+					switch {
+					case strings.HasPrefix(itemTrimArchSites, "password "):
+						var err error
+						sitesOptions["password"], err = jdecode.Decode(strings.Trim(strings.TrimPrefix(
+							itemTrimArchSites, "password "), "\""))
+						if err != nil {
+							return confRead, err
+						}
+					case strings.HasPrefix(itemTrimArchSites, "routing-instance "):
+						sitesOptions["routing_instance"] = strings.TrimPrefix(itemTrimArchSites, "routing-instance ")
+					}
+					archiveM["sites"] = append(archiveM["sites"].([]map[string]interface{}), sitesOptions)
+				}
+
+				// override (maxItem = 1)
+				confRead.archive = []map[string]interface{}{archiveM}
 			}
 		}
 	} else {
-		confRead.host = ""
+		confRead.filename = ""
 
 		return confRead, nil
 	}
@@ -553,10 +678,10 @@ func readSystemSyslogHost(host string, m interface{}, jnprSess *NetconfObject) (
 	return confRead, nil
 }
 
-func delSystemSyslogHost(host string, m interface{}, jnprSess *NetconfObject) error {
+func delSystemSyslogFile(filename string, m interface{}, jnprSess *NetconfObject) error {
 	sess := m.(*Session)
 	configSet := make([]string, 0, 1)
-	configSet = append(configSet, "delete system syslog host "+host)
+	configSet = append(configSet, "delete system syslog file "+filename)
 	err := sess.configSet(configSet, jnprSess)
 	if err != nil {
 		return err
@@ -564,108 +689,93 @@ func delSystemSyslogHost(host string, m interface{}, jnprSess *NetconfObject) er
 
 	return nil
 }
-func fillSystemSyslogHostData(d *schema.ResourceData, syslogHostOptions syslogHostOptions) {
-	tfErr := d.Set("host", syslogHostOptions.host)
+
+func fillSystemSyslogFileData(d *schema.ResourceData, syslogFileOptions syslogFileOptions) {
+	tfErr := d.Set("filename", syslogFileOptions.filename)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("allow_duplicates", syslogHostOptions.allowDuplicates)
+	tfErr = d.Set("allow_duplicates", syslogFileOptions.allowDuplicates)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("exclude_hostname", syslogHostOptions.excludeHostname)
+	tfErr = d.Set("explicit_priority", syslogFileOptions.explicitPriority)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("explicit_priority", syslogHostOptions.explicitPriority)
+	tfErr = d.Set("match", syslogFileOptions.match)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("facility_override", syslogHostOptions.facilityOverride)
+	tfErr = d.Set("match_strings", syslogFileOptions.matchStrings)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("log_prefix", syslogHostOptions.logPrefix)
+	tfErr = d.Set("structured_data", syslogFileOptions.structuredData)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("match", syslogHostOptions.match)
+	tfErr = d.Set("any_severity", syslogFileOptions.anySeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("match_strings", syslogHostOptions.matchStrings)
+	tfErr = d.Set("authorization_severity", syslogFileOptions.authorizationSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("port", syslogHostOptions.port)
+	tfErr = d.Set("changelog_severity", syslogFileOptions.changelogSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("source_address", syslogHostOptions.sourceAddress)
+	tfErr = d.Set("conflictlog_severity", syslogFileOptions.conflictlogSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("structured_data", syslogHostOptions.structuredData)
+	tfErr = d.Set("daemon_severity", syslogFileOptions.daemonSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("any_severity", syslogHostOptions.anySeverity)
+	tfErr = d.Set("dfc_severity", syslogFileOptions.dfcSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("authorization_severity", syslogHostOptions.authorizationSeverity)
+	tfErr = d.Set("external_severity", syslogFileOptions.externalSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("changelog_severity", syslogHostOptions.changelogSeverity)
+	tfErr = d.Set("firewall_severity", syslogFileOptions.firewallSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("conflictlog_severity", syslogHostOptions.conflictlogSeverity)
+	tfErr = d.Set("ftp_severity", syslogFileOptions.ftpSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("daemon_severity", syslogHostOptions.daemonSeverity)
+	tfErr = d.Set("interactivecommands_severity", syslogFileOptions.interactivecommandsSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("dfc_severity", syslogHostOptions.dfcSeverity)
+	tfErr = d.Set("kernel_severity", syslogFileOptions.kernelSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("external_severity", syslogHostOptions.externalSeverity)
+	tfErr = d.Set("ntp_severity", syslogFileOptions.ntpSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("firewall_severity", syslogHostOptions.firewallSeverity)
+	tfErr = d.Set("pfe_severity", syslogFileOptions.pfeSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("ftp_severity", syslogHostOptions.ftpSeverity)
+	tfErr = d.Set("security_severity", syslogFileOptions.securitySeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("interactivecommands_severity", syslogHostOptions.interactivecommandsSeverity)
+	tfErr = d.Set("user_severity", syslogFileOptions.userSeverity)
 	if tfErr != nil {
 		panic(tfErr)
 	}
-	tfErr = d.Set("kernel_severity", syslogHostOptions.kernelSeverity)
-	if tfErr != nil {
-		panic(tfErr)
-	}
-	tfErr = d.Set("ntp_severity", syslogHostOptions.ntpSeverity)
-	if tfErr != nil {
-		panic(tfErr)
-	}
-	tfErr = d.Set("pfe_severity", syslogHostOptions.pfeSeverity)
-	if tfErr != nil {
-		panic(tfErr)
-	}
-	tfErr = d.Set("security_severity", syslogHostOptions.securitySeverity)
-	if tfErr != nil {
-		panic(tfErr)
-	}
-	tfErr = d.Set("user_severity", syslogHostOptions.userSeverity)
+	tfErr = d.Set("archive", syslogFileOptions.archive)
 	if tfErr != nil {
 		panic(tfErr)
 	}